refactor: pluginify validator, save and new project (#227)
* feat(validator): move schema validator to plugin

* feat(validators/schemavalidator): instert button

* refactor(plugins): pluginify New and Save

* test(loaders/newproject-plugin): add integration tests

* test(validators/SchemaValidator): add integration test

* docs(hosting): fix copied Mixin doc comment
