--- test/unit/mock-validator.ts
@@ -1,5 +1,0 @@
-import { Validating } from '../../src/Validating.js';
-import { LitElement, customElement } from 'lit-element';
-
-@customElement('mock-validator')
-export class MockValidator extends Validating(LitElement) {}
